Move Exporter to its own class
