+++ script/benefit_application_issuers_export.rb
@@ -1,0 +1,1 @@
+BenefitSponsors::Exporters::BenefitApplicationIssuers.new.write
